Wrote a non-reactor, actor/producer/async-like workflow producer method and rewrote Reactor.toWorkflow to use it.
You can now write workflows imperatively, which is (probably?) a better fit for certain workflows. It is also
a much tighter scope of functionality to test, and the channel/job management logic isn't conflated with the
reactor loop logic.

The reactor logic is a lot simpler too (to read and to test), because it only contains the react loop, all the
channel management is factored out.

One further thing that could be done is to make `Workflow` extend the `Job` interface, might do this next.